Extract standard token behaviuor to reuse it in other tests
